add utils in C

+++ readFloat.h
@@ -1,10 +1,10 @@
 #include "readData.h"
 
-short* readInt16(const char* fn, size_t* outSize) {
+float* readFloat(const char* fn, size_t* outSize) {
   char* bin = readData(fn, outSize);
   if (bin == NULL) {
     return NULL;
   }
-  *outSize /= sizeof(short);
-  return (short*)bin;
+  *outSize /= sizeof(float);
+  return (float*)bin;
 }
